Move perRunCaches
